feat: progressbar onPushAudio, show audio time

+++ app/src/main/java/com/artf/chatapp/utils/extension/String.Ext.kt
@@ -5,6 +5,7 @@ import java.io.FileOutputStream
 import java.net.URL
 
 fun String.saveTo(path: String) {
+    if (File(path).exists()) return
     URL(this).openStream().use { input ->
         FileOutputStream(File(path)).use { output ->
             input.copyTo(output)
